+++ .gitignore
@@ -5,3 +5,4 @@ dashboard-audit.log
 dist/
 dashboard/dist/
 youtube-watcher2.js
+backup/
