+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "start": "node index.js",
-    "register-commands": "node register-commands.js"
+    "register-commands": "node src/utils/register-commands.js"
   },
   "engines": {
     "node": ">=16.9.0"
