+++ src/bot/commands/math.js
