+++ src/bot/commands/meme.js
