+++ src/bot/commands/poll.js
