+++ src/bot/commands/user.js
