+++ src/bot/commands/whoami.js
