+++ src/bot/services/auto-responses.js
