+++ src/utils/register-commands.js
